requirement.txt
opencv-python
facenet-pytorch
torch
torchvision
tqdm
pandas
numpy
scikit-learn
matplotlib
pillow
